Minor Fixes and Updates
I've add Cross Validatio to choose one of the parameters for the
Gradient Boosting model. that improves our score to 0.6805

+++ features/ad_count_per_display.py
@@ -6,7 +6,7 @@ def add_ad_count_per_display_feature(main_table):
     # Extracting the display_id's vector from main table
     displays_and_ads = main_table[["display_id"]]
 
-    # Counting how many times every display_id appears in the vector - this equlas the ad count per each display
+    # Counting how many times every display_id appears in the vector - this equals the ad count per each display
     ad_counts = displays_and_ads.groupby("display_id").display_id.agg("count")
     ad_counts = pd.Series(ad_counts.as_matrix())
 
@@ -17,6 +17,6 @@ def add_ad_count_per_display_feature(main_table):
     # (has only display_id and ad_count columns)
     res_frame = pd.DataFrame()
     res_frame["display_id"] = displays
-    res_frame["ad_count_per_display"] = ad_counts
+    res_frame["ad_count_per_display"] = ad_counts.astype(float)
 
     return res_frame
